+++ src/data/key/mod.rs
@@ -1,0 +1,33 @@
+// devela::data::key
+//
+//! Abstractions for managing unique associations and key-based storage.
+//!
+//! 
+//! Provides tools for mapping, indexing, and efficiently organizing data
+//! through unique keys, including maps, sets, and other key-value structures.
+//!
+#![doc = crate::doc_!(extends: collections)]
+//
+
+// mod map;
+// mod set;
+// mod trie;
+
+mod reexports;
+
+crate::items! { // structural access: _mods, _pub_mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+    #[allow(unused)] #[doc(hidden, no_inline)]
+    pub use _pub_mods::*;
+
+    mod _mods { #![allow(unused)]
+        pub use super::reexports::*;
+    }
+    mod _pub_mods { #![allow(unused)]
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::{_mods::*, _pub_mods::*};
+    }
+}