+++ src/data/list/link/mod.rs
@@ -11,7 +11,7 @@
 mod r#const;
 mod reexports;
 
-crate::items! { // structural access: _mods, _all
+crate::items! { // structural access: _mods, _all, _always
     #[allow(unused)]
     pub use _mods::*;
 
@@ -21,5 +21,8 @@ crate::items! { // structural access: _mods, _all
     pub(super) mod _all {
         #[doc(inline)]
         pub use super::_mods::*;
+    }
+    pub(super) mod _always { #![allow(unused)]
+        pub use super::{r#const::*, reexports::*};
     }
 }