+++ src/data/list/queue/mod.rs
@@ -15,7 +15,7 @@ mod adt;
 mod destaque;
 mod reexports;
 
-crate::items! { // structural access: _mods, _all
+crate::items! { // structural access: _mods, _all, _always
     #[allow(unused)]
     pub use _mods::*;
 
@@ -28,5 +28,8 @@ crate::items! { // structural access: _mods, _all
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
+    }
+    pub(super) mod _always { #![allow(unused)]
+        pub use super::{adt::*, reexports::*};
     }
 }