+++ src/data/list/queue/reexports.rs
@@ -3,9 +3,16 @@
 //! Reexported items.
 //
 
-crate::impl_cdef![<T> Self::new() => VecDeque<T>]; // impl ConstDefault
+use crate::{impl_cdef, reexport};
 
-crate::reexport! { rust: alloc::collections,
+impl_cdef![<T> Self::new() => VecDeque<T>]; // impl ConstDefault
+
+reexport! { rust: alloc::collections,
+    tag: crate::TAG_DATA_STRUCTURE!(),
+    doc: "A priority queue implemented with a binary heap.",
+    BinaryHeap
+}
+reexport! { rust: alloc::collections,
     tag: crate::TAG_DATA_STRUCTURE!(),
     doc: "A double-ended growable queue.",
     VecDeque