+++ src/data/node/mod.rs
@@ -6,17 +6,18 @@
 //! like linked lists, graphs and trees.
 //
 
-mod index;
-mod node;
+// mod index;
+// mod node;
 
 crate::items! { // structural access: _mods, _all
     #[allow(unused)]
     pub use _mods::*;
 
-    mod _mods {
-        pub use super::{index::*, node::*};
+    mod _mods { #![allow(unused)]
+        // pub use super::index::*; // WIP
+        // pub use super::node::*; // WIP
     }
-    pub(super) mod _all {
+    pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
     }